+++ apps/backend/src/modules/routes/route.controller.test.ts
@@ -102,6 +102,16 @@ describe("Route controller", () => {
 		duration: 4.56,
 		geometry: mockGeometry,
 		id: FIRST_POI_ID,
+		images: [
+			{
+				id: 1,
+				url: "https://s3.amazonaws.com/test/1.png",
+			},
+			{
+				id: 2,
+				url: "https://s3.amazonaws.com/test/2.png",
+			},
+		],
 		name: "Test Route",
 		pois: [
 			{ id: FIRST_POI_ID, name: FIRST_NAME, visitOrder: FIRST_VISIT_ORDER },