+++ apps/backend/src/modules/routes/route.model.ts
@@ -1,9 +1,12 @@
-import { Model, type RelationMappings } from "objection";
+import { type RouteUploadImageResponseDto } from "@smartscapes/shared";
+import { Model, type QueryBuilder, type RelationMappings } from "objection";
 
+import { FileFolderName } from "~/libs/enums/enums.js";
 import { DatabaseTableName } from "~/libs/modules/database/database.js";
 import { type LineStringGeometry } from "~/libs/types/types.js";
 
 import { CategoryModel } from "../categories/category.model.js";
+import { FileModel } from "../files/files.model.js";
 import { PointsOfInterestModel } from "../points-of-interest/points-of-interest.model.js";
 
 class RouteModel extends Model {
@@ -22,6 +25,8 @@ class RouteModel extends Model {
 	public geometry!: LineStringGeometry;
 
 	public id!: number;
+
+	public images!: RouteUploadImageResponseDto[];
 
 	public name!: string;
 
@@ -43,6 +48,17 @@ class RouteModel extends Model {
 			},
 			modelClass: CategoryModel,
 			relation: Model.ManyToManyRelation,
+		},
+		images: {
+			filter: (query: QueryBuilder<FileModel>): void => {
+				query.where("folder", FileFolderName.ROUTES);
+			},
+			join: {
+				from: "routes.id",
+				to: "files.entityId",
+			},
+			modelClass: FileModel,
+			relation: this.HasManyRelation,
 		},
 		pois: {
 			join: {