+++ apps/frontend/src/libs/modules/store/libs/types/root-reducer.type.ts
@@ -2,6 +2,7 @@ import { type reducer as appReducer } from "~/modules/app/app.js";
 import { type reducer as authReducer } from "~/modules/auth/auth.js";
 import { type reducer as exploreReducer } from "~/modules/explore/explore.js";
 import { type reducer as locationReducer } from "~/modules/location/location.js";
+import { type reducer as notificationReducer } from "~/modules/notification/notification.js";
 import {
 	type pointOfInterestDetailsReducer,
 	type pointsOfInterestReducer,
@@ -19,6 +20,7 @@ type RootReducer = {
 	constructRoute: ReturnType<typeof constructRouteReducer>;
 	explore: ReturnType<typeof exploreReducer>;
 	location: ReturnType<typeof locationReducer>;
+	notification: ReturnType<typeof notificationReducer>;
 	pointOfInterestDetails: ReturnType<typeof pointOfInterestDetailsReducer>;
 	pointsOfInterest: ReturnType<typeof pointsOfInterestReducer>;
 	routeDetails: ReturnType<typeof routeDetailsReducer>;