+++ apps/frontend/src/libs/modules/store/store.module.ts
@@ -14,6 +14,10 @@ import { authApi, reducer as authReducer } from "~/modules/auth/auth.js";
 import { reducer as exploreReducer } from "~/modules/explore/explore.js";
 import { reducer as locationReducer } from "~/modules/location/location.js";
 import {
+	notificationApi,
+	reducer as notificationReducer,
+} from "~/modules/notification/notification.js";
+import {
 	pointOfInterestApi,
 	pointOfInterestDetailsReducer,
 	pointsOfInterestReducer,
@@ -42,6 +46,7 @@ class Store {
 	public get extraArguments(): ExtraArguments {
 		return {
 			authApi,
+			notificationApi,
 			pointOfInterestApi,
 			reviewApi,
 			routesApi,
@@ -67,6 +72,7 @@ class Store {
 				constructRoute: constructRouteReducer,
 				explore: exploreReducer,
 				location: locationReducer,
+				notification: notificationReducer,
 				pointOfInterestDetails: pointOfInterestDetailsReducer,
 				pointsOfInterest: pointsOfInterestReducer,
 				routeDetails: routeDetailsReducer,