+++ packages/shared/src/modules/files/libs/enums/files-api-path.enum.ts
@@ -1,4 +1,5 @@
 const FilesApiPath = {
+	$ID: "/:id",
 	ROOT: "/",
 	UPLOAD: "/upload/:folder/:entityId",
 } as const;