+++ CHANGELOG.md
@@ -1,5 +1,42 @@
 # Changelog
 
+## [1.57.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.56.0...v1.57.0) (2025-08-26)
+
+
+### Features
+
+* **backend/frontend/shared:** implement edit and delete actions to pois ss-495 ([#543](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/543)) ([9c660ac](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/9c660ac75984b3c592339049f8dfa0cace25b078))
+
+
+### Bug Fixes
+
+* **frontend:** enhance routes pagination & input behavior ss-469 ([#492](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/492)) ([0a33a06](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/0a33a06af479b327a5f7b1c65227ecab836680e3))
+
+## [1.56.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.55.0...v1.56.0) (2025-08-25)
+
+
+### Features
+
+* **frontend/backend:** ability to upload images to the route page ss-476 ([#513](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/513)) ([e4fa4eb](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/e4fa4ebad3fbdee53a403e1d157bfeaa554e1d03))
+
+
+### Bug Fixes
+
+* **backend:** fix user routes request ss-534 ([#548](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/548)) ([69123cd](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/69123cd01c58783fb15a2678eeeded290f74a0de))
+
+## [1.55.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.54.0...v1.55.0) (2025-08-25)
+
+
+### Features
+
+* **frontend/backend:** create and display follow notification ss-438 ([#522](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/522)) ([49f60f6](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/49f60f6a3b4fa87da3525311909c1fd6526b5e12))
+
+
+### Bug Fixes
+
+* **backend:** fix user routes request ss-534 ([#538](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/538)) ([86b0af4](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/86b0af4876e43d79ec249f221e5555f58c00d427))
+* **frontend:** add styles to prevent long review overflowing ss-540 ([#541](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/541)) ([aa19db9](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/aa19db9fdddb1949515afc8ea47d17b5ec1b94a7))
+
 ## [1.54.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.53.0...v1.54.0) (2025-08-25)